AB#3115 Moving project field commands to `project field`
