+++ austrakka/components/project/funcs.py
@@ -1,13 +1,8 @@
-from typing import List
-
-import pandas as pd
-
 from austrakka.utils.api import call_api
+from austrakka.utils.api import post
 from austrakka.utils.helpers.output import call_get_and_print_table
-from austrakka.utils.api import post, patch, get
 from austrakka.utils.misc import logger_wraps
 from austrakka.utils.paths import PROJECT_PATH
-from austrakka.utils.output import print_table
 
 
 @logger_wraps()
@@ -24,59 +19,6 @@ def add_project(abbrev: str, name: str, description: str, org: str):
             },
             "isActive": True
         }
-    )
-
-
-@logger_wraps()
-def set_field_project(abbrev: str, field_names: List[str]):
-
-    field_name_objs = []
-    for field_name in field_names:
-        field_name_objs.append({
-            "ColumnName": field_name
-        })
-
-    return call_api(
-        method=patch,
-        path=f'{PROJECT_PATH}/set-fields/{abbrev}',
-        body={
-            "columnNames": field_name_objs
-        }
-    )
-
-
-@logger_wraps()
-def unset_field_project(abbrev: str, field_names: List[str]):
-
-    field_name_objs = []
-    for field_name in field_names:
-        field_name_objs.append({
-            "ColumnName": field_name
-        })
-
-    return call_api(
-        method=patch,
-        path=f'{PROJECT_PATH}/unset-fields/{abbrev}',
-        body={
-            "columnNames": field_name_objs
-        }
-    )
-
-
-@logger_wraps()
-def display_field_project(abbrev: str, table_format):
-
-    response = call_api(
-        method=get,
-        path=f'{PROJECT_PATH}/display-fields/{abbrev}',
-    )
-
-    data = response['data'] if ('data' in response) else response
-    result = pd.json_normalize(data, max_level=1)
-
-    print_table(
-        result,
-        table_format,
     )
